refactor: move StringEncoding to module
