+++ read_buffer/src/column/encoding/dictionary/rle.rs
@@ -579,6 +579,24 @@ impl RLE {
         max
     }
 
+    /// Returns the lexicographical minimum value in the column. None is
+    /// returned only if the column does not contain any non-null values.
+    pub fn column_min(&self) -> Option<&'_ String> {
+        if self.index_entries.len() > 1 {
+            return Some(self.index_entries.get(1).unwrap());
+        }
+        None
+    }
+
+    /// Returns the lexicographical maximum value in the column. None is
+    /// returned only if the column does not contain any non-null values.
+    pub fn column_max(&self) -> Option<&'_ String> {
+        if self.index_entries.len() > 1 {
+            return Some(self.index_entries.last().unwrap());
+        }
+        None
+    }
+
     /// Returns the total number of non-null values found at the provided set of
     /// row ids.
     pub fn count(&self, row_ids: &[u32]) -> u32 {
